Removed all dependecy to the apartments.js files
