+++ client/package.json
@@ -20,7 +20,7 @@
     "react-router-bootstrap": "^0.24.2",
     "react-router-dom": "^4.1.2",
     "react-router-redux": "^4.0.8",
-    "react-scripts": "1.0.5",
+    "react-scripts": "1.0.13",
     "redux": "^3.7.1",
     "redux-auth": "^0.0.5-beta5",
     "redux-logger": "^3.0.6",
