+++ client/src/containers/Home.js
@@ -4,7 +4,7 @@ import PropTypes from 'prop-types';
 import HeroCTA from '../components/HeroCTA';
 import ApartmentMap from '../components/CampaignsMap';
 
-import { fetchApartmentsRequest } from '../redux/actions/apartments';
+import { fetchApartmentsRequest } from '../redux/actions/initialSearch';
 import { openMap, closeMap } from '../redux/actions/googleMap';
 
 class Home extends Component {
@@ -13,7 +13,7 @@ class Home extends Component {
   }
 
   render() {
-    const { apartments: { apartments }, googleMap: { isOpen } } = this.props;
+    const { initialSearch: { apartments }, googleMap: { isOpen } } = this.props;
     return (
       <div>
         <HeroCTA
@@ -31,7 +31,7 @@ class Home extends Component {
 
 Home.propTypes = {
   fetchApartmentsRequest: PropTypes.func.isRequired,
-  apartments: PropTypes.shape({
+  initialSearch: PropTypes.shape({
     apartments: PropTypes.array.isRequired
   }).isRequired,
   googleMap: PropTypes.shape({
@@ -42,5 +42,5 @@ Home.propTypes = {
 };
 
 export default connect(
-  ({ apartments, googleMap }) => ({ apartments, googleMap }),
+  ({ initialSearch, googleMap }) => ({ initialSearch, googleMap }),
   { fetchApartmentsRequest, openMap, closeMap })(Home);
