--- client/src/redux/constants/apartments.js
@@ -1,3 +1,0 @@
-export const APARTMENTS_REQUEST = 'APARTMENTS_REQUEST';
-export const APARTMENTS_SUCCESS = 'APARTMENTS_SUCCESS';
-export const APARTMENTS_FAILURE = 'APARTMENTS_FAILURE';
