+++ client/src/redux/constants/initialSearch.js
@@ -10,3 +10,7 @@ export const CLEAR_SEARCH_RESULTS = 'CLEAR_SEARCH_RESULTS';
 
 export const SET_CAMPAIGN_INFORMATION = 'SET_CAMPAIGN_INFORMATION';
 export const SELECT_ADDRESS = 'SELECT_ADDRESS';
+
+export const APARTMENTS_REQUEST = 'APARTMENTS_REQUEST';
+export const APARTMENTS_SUCCESS = 'APARTMENTS_SUCCESS';
+export const APARTMENTS_FAILURE = 'APARTMENTS_FAILURE';
