+++ client/src/redux/reducers/initialSearch.js
@@ -6,7 +6,10 @@ import {
   VALIDATE_ADDRESS_FAILURE,
   CLEAR_SEARCH_RESULTS,
   SET_CAMPAIGN_INFORMATION,
-  SELECT_ADDRESS
+  SELECT_ADDRESS,
+  APARTMENTS_REQUEST,
+  APARTMENTS_SUCCESS,
+  APARTMENTS_FAILURE
 } from '../constants/initialSearch';
 
 const defaultState = {
@@ -78,6 +81,26 @@ export default function (state = defaultState, action) {
       return {
         ...state,
         campaignInfo: action.campaignInfo
+      };
+    case APARTMENTS_REQUEST:
+      return {
+        ...state,
+        loading: true,
+        loaded: false
+      };
+    case APARTMENTS_SUCCESS:
+      return {
+        ...state,
+        loading: false,
+        loaded: true,
+        apartments: response
+      };
+    case APARTMENTS_FAILURE:
+      return {
+        ...state,
+        loading: false,
+        loaded: false,
+        error
       };
     default:
       return state;
